Rename Service namespace to Container
see: https://github.com/prooph/event-store/issues/69
